\name{hierNetGxE-package}
\alias{hierNetGxE-package}
\alias{hierNetGxE}
\docType{package}
\title{
\packageTitle{hierNetGxE}
}
\description{
\packageDescription{hierNetGxE}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{hierNetGxE}
\packageIndices{hierNetGxE}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{hierNetGxE}

Maintainer: \packageMaintainer{hierNetGxE}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
