\name{gesso-package}
\alias{gesso-package}
\alias{gesso}
\docType{package}
\title{
\packageTitle{gesso}
}
\description{
\packageDescription{gesso}
}
\author{
\packageAuthor{gesso}

Maintainer: \packageMaintainer{gesso}
}
\references{
"A Scalable Hierarchical Lasso for Gene-Environment Interactions", 
Natalia Zemlianskaia, W.James Gauderman, Juan Pablo Lewinger
https://arxiv.org/abs/2103.13510
}
\keyword{ package }


