\name{gesso-package}
\alias{gesso-package}
\alias{gesso}
\docType{package}
\title{
\packageTitle{gesso}
}
\description{
\packageDescription{gesso}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{gesso}
\packageIndices{gesso}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{gesso}

Maintainer: \packageMaintainer{gesso}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
